+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - support for WCAG 1.3.1 - Info and Relationships
 - `autocompleteAttributes` to input, options and date based elements
+- support for WCAG 4.1.3 - Status Messages
 
 ### Changed